+++ src/main/kotlin/net/internetisalie/lunar/lang/completion/postfix/LuaForIpairsPostfixTemplate.kt
@@ -4,14 +4,14 @@ import com.intellij.codeInsight.template.postfix.templates.StringBasedPostfixTem
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateProvider
 import com.intellij.psi.PsiElement
 
-class LuaIfPostfixTemplate(provider: PostfixTemplateProvider? = null) : StringBasedPostfixTemplate(
-    "if",
-    "if expr then ... end",
+class LuaForIpairsPostfixTemplate(provider: PostfixTemplateProvider? = null) : StringBasedPostfixTemplate(
+    "fori",
+    "for i, v in ipairs(expr) do ... end",
     LuaExprSelector(),
     provider
 ) {
     override fun getTemplateString(element: PsiElement): String {
-        return "if \$expr\$ then\n    \$END\$\nend"
+        return "for i, v in ipairs(\$expr\$) do\n    \$END\$\nend"
     }
 
     override fun getElementToRemove(expr: PsiElement): PsiElement {